backend/requirements/base.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Computer Vision (from existing requirements)
opencv-python==4.9.0.80
ultralytics==8.1.0
# Note: dlib will be installed from the .whl file

# Telegram
python-telegram-bot==20.7

# Logging
loguru==0.7.2

# Image processing
Pillow==10.2.0
numpy==1.26.3
